Remove classes deprecated in 2020-2022.
