+++ integration-tests/kubernetes-client/src/test/java/io/quarkus/it/kubernetes/client/AbsentConfigMapPropertiesPMT.java
@@ -1,7 +1,5 @@
 package io.quarkus.it.kubernetes.client;
 
-import java.io.IOException;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -9,7 +7,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import io.quarkus.test.QuarkusProdModeTest;
 import io.quarkus.test.common.QuarkusTestResource;
 
-@QuarkusTestResource(value = CustomKubernetesMockServerTestResource.class, restrictToAnnotatedClass = true)
+@QuarkusTestResource(value = OtherCustomKubernetesServerTestResource.class, restrictToAnnotatedClass = true)
 public class AbsentConfigMapPropertiesPMT {
 
     @RegisterExtension
@@ -22,7 +20,7 @@ public class AbsentConfigMapPropertiesPMT {
             .setApplicationVersion("0.1-SNAPSHOT");
 
     @Test
-    public void startUpShouldFail() throws IOException {
+    public void startUpShouldFail() {
         Assertions.assertThat(config.getStartupConsoleOutput())
                 .contains("ConfigMap 'cmap4' not found in namespace 'demo'").contains("RuntimeException");
     }
