+++ integration-tests/kubernetes-client/src/test/java/io/quarkus/it/kubernetes/client/ConfigMapPropertiesTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-@QuarkusTestResource(value = CustomKubernetesMockServerTestResource.class, restrictToAnnotatedClass = true)
+@QuarkusTestResource(value = OtherCustomKubernetesServerTestResource.class, restrictToAnnotatedClass = true)
 @QuarkusTest
 public class ConfigMapPropertiesTest {
