--- integration-tests/kubernetes-client/src/test/java/io/quarkus/it/kubernetes/client/KubernetesNewClientTestIT.java
@@ -1,8 +1,0 @@
-package io.quarkus.it.kubernetes.client;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class KubernetesNewClientTestIT extends KubernetesNewClientTest {
-
-}
